Clean up debug implementation
Keeps only the records based variant of attaching the
debugger for now.
