+++ lib/js/src/tea_debug.js
@@ -10,14 +10,20 @@ var $$String = require("bs-platform/lib/js/string.js");
 var Tea_app = require("./tea_app.js");
 var Tea_cmd = require("./tea_cmd.js");
 var Tea_sub = require("./tea_sub.js");
-var Tea_html = require("./tea_html.js");
+var Tea_html2 = require("./tea_html2.js");
 
-function MakeStandardProgram(Client) {
-  var client_msg = function (msg) {
-    return /* ClientMsg */Block.__(0, [msg]);
-  };
-  var init = function (flags) {
-    var match = Curry._1(Client[/* init */0], flags);
+function client_msg(msg) {
+  return /* ClientMsg */Block.__(0, [msg]);
+}
+
+function debug(string_of_msg, param) {
+  var shutdown = param[/* shutdown */4];
+  var subscriptions = param[/* subscriptions */3];
+  var view = param[/* view */2];
+  var update = param[/* update */1];
+  var init = param[/* init */0];
+  var init$prime = function (flags) {
+    var match = Curry._1(init, flags);
     return /* tuple */[
             /* record */[
               /* history : :: */[
@@ -33,7 +39,7 @@ function MakeStandardProgram(Client) {
             Tea_cmd.map(client_msg, match[1])
           ];
   };
-  var update = function (model, param) {
+  var update$prime = function (model, param) {
     if (typeof param === "number") {
       if (param === 0) {
         var match = model[/* state */1];
@@ -78,10 +84,10 @@ function MakeStandardProgram(Client) {
     } else if (model[/* state */1] === /* Running */0) {
       var msg = param[0];
       var match$1 = List.hd(model[/* history */0]);
-      var match$2 = Curry._2(Client[/* update */1], match$1[1], msg);
+      var match$2 = Curry._2(update, match$1[1], msg);
       var dmodel$prime_000 = /* history : :: */[
         /* tuple */[
-          Curry._1(Client[/* string_of_msg */4], msg),
+          Curry._1(string_of_msg, msg),
           match$2[0]
         ],
         model[/* history */0]
@@ -125,7 +131,7 @@ function MakeStandardProgram(Client) {
                                 return param[0] + (":" + param[1]);
                               }), properties)))]);
     };
-    return Tea_html.node(/* None */0, "style", /* None */0, /* None */0, /* [] */0, /* :: */[
+    return Tea_html2.node(/* None */0, "style", /* None */0, /* None */0, /* [] */0, /* :: */[
                 rule("#debug.paused", /* :: */[
                       /* tuple */[
                         "position",
@@ -666,13 +672,10 @@ function MakeStandardProgram(Client) {
   };
   var view_details = function (model) {
     var format = (
-      function (v) { return JSON.stringify(v, null, 2); }
-    );
-    return Tea_html.aside(/* None */0, /* None */0, /* :: */[
-                /* RawProp */Block.__(0, [
-                    "className",
-                    "details"
-                  ]),
+        function (v) { return JSON.stringify(v, null, 2); }
+      );
+    return Tea_html2.aside(/* None */0, /* None */0, /* :: */[
+                Curry._1(Tea_html2.Attributes[/* class' */3], "details"),
                 /* [] */0
               ], /* :: */[
                 /* Text */Block.__(1, [Curry._1(format, model)]),
@@ -681,19 +684,15 @@ function MakeStandardProgram(Client) {
   };
   var view_history = function (model, selected_index) {
     var count = List.length(model[/* history */0]);
-    return Tea_html.ul(/* None */0, /* None */0, /* :: */[
-                /* RawProp */Block.__(0, [
-                    "className",
-                    "history"
-                  ]),
+    return Tea_html2.ul(/* None */0, /* None */0, /* :: */[
+                Curry._1(Tea_html2.Attributes[/* class' */3], "history"),
                 /* [] */0
               ], List.mapi((function (i, param) {
                       var selected = +(i === selected_index);
-                      var msg = /* SelectHistoryItem */Block.__(1, [i]);
-                      return Tea_html.li(/* None */0, /* None */0, /* :: */[
-                                  Vdom.onMsg("click", msg),
+                      return Tea_html2.li(/* None */0, /* None */0, /* :: */[
+                                  Curry._1(Tea_html2.Events[/* onClick */8], /* SelectHistoryItem */Block.__(1, [i])),
                                   /* :: */[
-                                    Tea_html.classList(/* :: */[
+                                    Curry._1(Tea_html2.Attributes[/* classList */4], /* :: */[
                                           /* tuple */[
                                             "selected",
                                             selected
@@ -703,8 +702,8 @@ function MakeStandardProgram(Client) {
                                     /* [] */0
                                   ]
                                 ], /* :: */[
-                                  Tea_html.span(/* None */0, /* None */0, /* :: */[
-                                        Tea_html.classList(/* :: */[
+                                  Tea_html2.span(/* None */0, /* None */0, /* :: */[
+                                        Curry._1(Tea_html2.Attributes[/* classList */4], /* :: */[
                                               /* tuple */[
                                                 "details",
                                                 /* true */1
@@ -718,43 +717,33 @@ function MakeStandardProgram(Client) {
                                               ]
                                             ]),
                                         selected ? /* :: */[
-                                            Vdom.onMsg("click", /* ToggleDetails */1),
+                                            Curry._1(Tea_html2.Events[/* onClick */8], /* ToggleDetails */1),
                                             /* :: */[
-                                              /* Attribute */Block.__(1, [
-                                                  "",
-                                                  "title",
-                                                  "toggle details"
-                                                ]),
+                                              Curry._1(Tea_html2.Attributes[/* title */6], "toggle details"),
                                               /* [] */0
                                             ]
                                           ] : /* :: */[
-                                            Tea_html.noProp,
+                                            Tea_html2.Attributes[/* noProp */0],
                                             /* :: */[
-                                              Tea_html.noProp,
+                                              Tea_html2.Attributes[/* noProp */0],
                                               /* [] */0
                                             ]
                                           ]
                                       ], /* :: */[
-                                        selected && model[/* show_details */2] ? view_details(param[1]) : Tea_html.noNode,
+                                        selected && model[/* show_details */2] ? view_details(param[1]) : Tea_html2.noNode,
                                         /* [] */0
                                       ]),
                                   /* :: */[
-                                    Tea_html.span(/* None */0, /* None */0, /* :: */[
-                                          /* RawProp */Block.__(0, [
-                                              "className",
-                                              "message"
-                                            ]),
+                                    Tea_html2.span(/* None */0, /* None */0, /* :: */[
+                                          Curry._1(Tea_html2.Attributes[/* class' */3], "message"),
                                           /* [] */0
                                         ], /* :: */[
                                           /* Text */Block.__(1, [param[0]]),
                                           /* [] */0
                                         ]),
                                     /* :: */[
-                                      Tea_html.span(/* None */0, /* None */0, /* :: */[
-                                            /* RawProp */Block.__(0, [
-                                                "className",
-                                                "index"
-                                              ]),
+                                      Tea_html2.span(/* None */0, /* None */0, /* :: */[
+                                            Curry._1(Tea_html2.Attributes[/* class' */3], "index"),
                                             /* [] */0
                                           ], /* :: */[
                                             /* Text */Block.__(1, [String(count - i | 0)]),
@@ -766,7 +755,7 @@ function MakeStandardProgram(Client) {
                                 ]);
                     }), model[/* history */0]));
   };
-  var view = function (model) {
+  var view$prime = function (model) {
     var match = model[/* state */1];
     var match$1;
     if (match) {
@@ -785,17 +774,14 @@ function MakeStandardProgram(Client) {
     }
     var paused = match$1[2];
     var history_count = List.length(model[/* history */0]);
-    var vnode = Curry._1(Client[/* view */2], match$1[1]);
-    return Tea_html.div(/* None */0, /* None */0, /* [] */0, /* :: */[
+    var vnode = Curry._1(view, match$1[1]);
+    return Tea_html2.div(/* None */0, /* None */0, /* [] */0, /* :: */[
                 Vdom.map(client_msg, vnode),
                 /* :: */[
-                  Tea_html.div(/* None */0, /* None */0, /* :: */[
-                        /* RawProp */Block.__(0, [
-                            "id",
-                            "debug"
-                          ]),
+                  Tea_html2.div(/* None */0, /* None */0, /* :: */[
+                        Curry._1(Tea_html2.Attributes[/* id */5], "debug"),
                         /* :: */[
-                          Tea_html.classList(/* :: */[
+                          Curry._1(Tea_html2.Attributes[/* classList */4], /* :: */[
                                 /* tuple */[
                                   "paused",
                                   paused
@@ -807,24 +793,13 @@ function MakeStandardProgram(Client) {
                       ], /* :: */[
                         view_styles(/* () */0),
                         /* :: */[
-                          Tea_html.nav(/* None */0, /* None */0, /* [] */0, /* :: */[
-                                Tea_html.div(/* None */0, /* None */0, /* :: */[
-                                      /* RawProp */Block.__(0, [
-                                          "className",
-                                          "toggle"
-                                        ]),
+                          Tea_html2.nav(/* None */0, /* None */0, /* [] */0, /* :: */[
+                                Tea_html2.div(/* None */0, /* None */0, /* :: */[
+                                      Curry._1(Tea_html2.Attributes[/* class' */3], "toggle"),
                                       /* :: */[
-                                        Vdom.onMsg("click", /* TogglePaused */0),
+                                        Curry._1(Tea_html2.Events[/* onClick */8], /* TogglePaused */0),
                                         /* :: */[
-                                          paused ? /* Attribute */Block.__(1, [
-                                                "",
-                                                "title",
-                                                "click to resume"
-                                              ]) : /* Attribute */Block.__(1, [
-                                                "",
-                                                "title",
-                                                "click to pause"
-                                              ]),
+                                          paused ? Curry._1(Tea_html2.Attributes[/* title */6], "click to resume") : Curry._1(Tea_html2.Attributes[/* title */6], "click to pause"),
                                           /* [] */0
                                         ]
                                       ]
@@ -847,7 +822,7 @@ function MakeStandardProgram(Client) {
                                       /* [] */0
                                     ]),
                                 /* :: */[
-                                  paused ? view_history(model, match$1[0]) : Tea_html.noNode,
+                                  paused ? view_history(model, match$1[0]) : Tea_html2.noNode,
                                   /* [] */0
                                 ]
                               ]),
@@ -858,21 +833,74 @@ function MakeStandardProgram(Client) {
                 ]
               ]);
   };
-  var subscriptions = function (model) {
-    var subs = Curry._1(Client[/* subscriptions */3], List.hd(model[/* history */0])[1]);
-    return Tea_sub.map(client_msg, subs);
+  var subscriptions$prime = function (model) {
+    return Tea_sub.map(client_msg, Curry._1(subscriptions, List.hd(model[/* history */0])[1]));
   };
-  var partial_arg = /* record */[
-    /* init */init,
-    /* update */update,
-    /* view */view,
-    /* subscriptions */subscriptions
-  ];
-  var start = function (param, param$1) {
-    return Tea_app.standardProgram(partial_arg, param, param$1);
+  var shutdown$prime = function (model) {
+    return Tea_cmd.map(client_msg, Curry._1(shutdown, List.hd(model[/* history */0])[1]));
   };
-  return /* module */[/* start */start];
+  return /* record */[
+          /* init */init$prime,
+          /* update */update$prime,
+          /* view */view$prime,
+          /* subscriptions */subscriptions$prime,
+          /* shutdown */shutdown$prime
+        ];
 }
 
-exports.MakeStandardProgram = MakeStandardProgram;
-/* Tea_html Not a pure module */
+function beginnerProgram(param, string_of_msg, pnode, flags) {
+  var update = param[/* update */1];
+  var model = param[/* model */0];
+  var debugged = debug(string_of_msg, /* record */[
+        /* init */(function () {
+            return /* tuple */[
+                    model,
+                    /* NoCmd */0
+                  ];
+          }),
+        /* update */(function (model, msg) {
+            return /* tuple */[
+                    Curry._2(update, model, msg),
+                    /* NoCmd */0
+                  ];
+          }),
+        /* view */param[/* view */2],
+        /* subscriptions */(function () {
+            return /* NoSub */0;
+          }),
+        /* shutdown */(function () {
+            return /* NoCmd */0;
+          })
+      ]);
+  return Tea_app.program(debugged, pnode, flags);
+}
+
+function standardProgram(param, string_of_msg, pnode, flags) {
+  var debugged = debug(string_of_msg, /* record */[
+        /* init */param[/* init */0],
+        /* update */param[/* update */1],
+        /* view */param[/* view */2],
+        /* subscriptions */param[/* subscriptions */3],
+        /* shutdown */(function () {
+            return /* NoCmd */0;
+          })
+      ]);
+  return Tea_app.program(debugged, pnode, flags);
+}
+
+function program(param, string_of_msg, pnode, flags) {
+  var debugged = debug(string_of_msg, /* record */[
+        /* init */param[/* init */0],
+        /* update */param[/* update */1],
+        /* view */param[/* view */2],
+        /* subscriptions */param[/* subscriptions */3],
+        /* shutdown */param[/* shutdown */4]
+      ]);
+  return Tea_app.program(debugged, pnode, flags);
+}
+
+exports.debug = debug;
+exports.beginnerProgram = beginnerProgram;
+exports.standardProgram = standardProgram;
+exports.program = program;
+/* Tea_html2 Not a pure module */
