--- lib/js/src/tea_debug_records.js
@@ -1,853 +1,0 @@
-// Generated by BUCKLESCRIPT VERSION 2.2.3, PLEASE EDIT WITH CARE
-'use strict';
-
-var List = require("bs-platform/lib/js/list.js");
-var Vdom = require("./vdom.js");
-var Block = require("bs-platform/lib/js/block.js");
-var Curry = require("bs-platform/lib/js/curry.js");
-var Printf = require("bs-platform/lib/js/printf.js");
-var $$String = require("bs-platform/lib/js/string.js");
-var Tea_app = require("./tea_app.js");
-var Tea_cmd = require("./tea_cmd.js");
-var Tea_sub = require("./tea_sub.js");
-var Tea_html2 = require("./tea_html2.js");
-
-function client_msg(msg) {
-  return /* ClientMsg */Block.__(0, [msg]);
-}
-
-function wrap(param) {
-  var string_of_msg = param[/* string_of_msg */4];
-  var subscriptions = param[/* subscriptions */3];
-  var view = param[/* view */2];
-  var update = param[/* update */1];
-  var init = param[/* init */0];
-  var init$prime = function (flags) {
-    var match = Curry._1(init, flags);
-    return /* tuple */[
-            /* record */[
-              /* history : :: */[
-                /* tuple */[
-                  "_init_",
-                  match[0]
-                ],
-                /* [] */0
-              ],
-              /* state : Running */0,
-              /* show_details : false */0
-            ],
-            Tea_cmd.map(client_msg, match[1])
-          ];
-  };
-  var update$prime = function (model, param) {
-    if (typeof param === "number") {
-      if (param === 0) {
-        var match = model[/* state */1];
-        if (match) {
-          return /* tuple */[
-                  /* record */[
-                    /* history */model[/* history */0],
-                    /* state : Running */0,
-                    /* show_details */model[/* show_details */2]
-                  ],
-                  /* NoCmd */0
-                ];
-        } else {
-          return /* tuple */[
-                  /* record */[
-                    /* history */model[/* history */0],
-                    /* state : Paused */[0],
-                    /* show_details */model[/* show_details */2]
-                  ],
-                  /* NoCmd */0
-                ];
-        }
-      } else {
-        return /* tuple */[
-                /* record */[
-                  /* history */model[/* history */0],
-                  /* state */model[/* state */1],
-                  /* show_details */1 - model[/* show_details */2]
-                ],
-                /* NoCmd */0
-              ];
-      }
-    } else if (param.tag) {
-      return /* tuple */[
-              /* record */[
-                /* history */model[/* history */0],
-                /* state : Paused */[param[0]],
-                /* show_details */model[/* show_details */2]
-              ],
-              /* NoCmd */0
-            ];
-    } else if (model[/* state */1] === /* Running */0) {
-      var msg = param[0];
-      var match$1 = List.hd(model[/* history */0]);
-      var match$2 = Curry._2(update, match$1[1], msg);
-      var dmodel$prime_000 = /* history : :: */[
-        /* tuple */[
-          Curry._1(string_of_msg, msg),
-          match$2[0]
-        ],
-        model[/* history */0]
-      ];
-      var dmodel$prime_001 = /* state */model[/* state */1];
-      var dmodel$prime_002 = /* show_details */model[/* show_details */2];
-      var dmodel$prime = /* record */[
-        dmodel$prime_000,
-        dmodel$prime_001,
-        dmodel$prime_002
-      ];
-      return /* tuple */[
-              dmodel$prime,
-              Tea_cmd.map(client_msg, match$2[1])
-            ];
-    } else {
-      return /* tuple */[
-              model,
-              /* NoCmd */0
-            ];
-    }
-  };
-  var view_styles = function () {
-    var rule = function (selector, properties) {
-      return /* Text */Block.__(1, [Curry._2(Printf.sprintf(/* Format */[
-                          /* String */Block.__(2, [
-                              /* No_padding */0,
-                              /* String_literal */Block.__(11, [
-                                  " {",
-                                  /* String */Block.__(2, [
-                                      /* No_padding */0,
-                                      /* Char_literal */Block.__(12, [
-                                          /* "}" */125,
-                                          /* End_of_format */0
-                                        ])
-                                    ])
-                                ])
-                            ]),
-                          "%s {%s}"
-                        ]), selector, $$String.concat(";", List.map((function (param) {
-                                return param[0] + (":" + param[1]);
-                              }), properties)))]);
-    };
-    return Tea_html2.node(/* None */0, "style", /* None */0, /* None */0, /* [] */0, /* :: */[
-                rule("#debug.paused", /* :: */[
-                      /* tuple */[
-                        "position",
-                        "fixed"
-                      ],
-                      /* :: */[
-                        /* tuple */[
-                          "top",
-                          "0"
-                        ],
-                        /* :: */[
-                          /* tuple */[
-                            "left",
-                            "0"
-                          ],
-                          /* :: */[
-                            /* tuple */[
-                              "width",
-                              "100%"
-                            ],
-                            /* :: */[
-                              /* tuple */[
-                                "height",
-                                "100%"
-                              ],
-                              /* :: */[
-                                /* tuple */[
-                                  "pointer-events",
-                                  "all"
-                                ],
-                                /* :: */[
-                                  /* tuple */[
-                                    "background-color",
-                                    "rgba(0,0,0,.1)"
-                                  ],
-                                  /* :: */[
-                                    /* tuple */[
-                                      "box-shadow",
-                                      "inset 0 0 10px #333"
-                                    ],
-                                    /* [] */0
-                                  ]
-                                ]
-                              ]
-                            ]
-                          ]
-                        ]
-                      ]
-                    ]),
-                /* :: */[
-                  rule("#debug nav", /* :: */[
-                        /* tuple */[
-                          "position",
-                          "fixed"
-                        ],
-                        /* :: */[
-                          /* tuple */[
-                            "bottom",
-                            "0"
-                          ],
-                          /* :: */[
-                            /* tuple */[
-                              "right",
-                              "6px"
-                            ],
-                            /* :: */[
-                              /* tuple */[
-                                "border-radius",
-                                "4px 4px 0 0"
-                              ],
-                              /* :: */[
-                                /* tuple */[
-                                  "background-color",
-                                  "#444"
-                                ],
-                                /* :: */[
-                                  /* tuple */[
-                                    "color",
-                                    "#fff"
-                                  ],
-                                  /* :: */[
-                                    /* tuple */[
-                                      "font-family",
-                                      "monospace"
-                                    ],
-                                    /* :: */[
-                                      /* tuple */[
-                                        "box-shadow",
-                                        "0 0 10px #333"
-                                      ],
-                                      /* [] */0
-                                    ]
-                                  ]
-                                ]
-                              ]
-                            ]
-                          ]
-                        ]
-                      ]),
-                  /* :: */[
-                    rule("#debug.paused nav", /* :: */[
-                          /* tuple */[
-                            "height",
-                            "50%"
-                          ],
-                          /* :: */[
-                            /* tuple */[
-                              "padding-bottom",
-                              "2em"
-                            ],
-                            /* [] */0
-                          ]
-                        ]),
-                    /* :: */[
-                      rule("#debug nav .toggle", /* :: */[
-                            /* tuple */[
-                              "padding",
-                              "6px"
-                            ],
-                            /* :: */[
-                              /* tuple */[
-                                "cursor",
-                                "pointer"
-                              ],
-                              /* :: */[
-                                /* tuple */[
-                                  "min-width",
-                                  "24ch"
-                                ],
-                                /* :: */[
-                                  /* tuple */[
-                                    "text-align",
-                                    "center"
-                                  ],
-                                  /* [] */0
-                                ]
-                              ]
-                            ]
-                          ]),
-                      /* :: */[
-                        rule("#debug nav .toggle:before", /* :: */[
-                              /* tuple */[
-                                "content",
-                                "' '"
-                              ],
-                              /* :: */[
-                                /* tuple */[
-                                  "position",
-                                  "absolute"
-                                ],
-                                /* :: */[
-                                  /* tuple */[
-                                    "left",
-                                    "0"
-                                  ],
-                                  /* :: */[
-                                    /* tuple */[
-                                      "top",
-                                      "0"
-                                    ],
-                                    /* :: */[
-                                      /* tuple */[
-                                        "width",
-                                        ".5ch"
-                                      ],
-                                      /* :: */[
-                                        /* tuple */[
-                                          "height",
-                                          "1.8ch"
-                                        ],
-                                        /* :: */[
-                                          /* tuple */[
-                                            "margin",
-                                            ".8ch 1ch"
-                                          ],
-                                          /* :: */[
-                                            /* tuple */[
-                                              "border",
-                                              "solid #fff"
-                                            ],
-                                            /* :: */[
-                                              /* tuple */[
-                                                "border-width",
-                                                "0 .5ch"
-                                              ],
-                                              /* [] */0
-                                            ]
-                                          ]
-                                        ]
-                                      ]
-                                    ]
-                                  ]
-                                ]
-                              ]
-                            ]),
-                        /* :: */[
-                          rule("#debug.paused nav .toggle:before", /* :: */[
-                                /* tuple */[
-                                  "border-color",
-                                  "transparent"
-                                ],
-                                /* :: */[
-                                  /* tuple */[
-                                    "border-left-color",
-                                    "#fff"
-                                  ],
-                                  /* :: */[
-                                    /* tuple */[
-                                      "border-width",
-                                      "1ch"
-                                    ],
-                                    /* :: */[
-                                      /* tuple */[
-                                        "width",
-                                        "0"
-                                      ],
-                                      /* :: */[
-                                        /* tuple */[
-                                          "height",
-                                          "0"
-                                        ],
-                                        /* [] */0
-                                      ]
-                                    ]
-                                  ]
-                                ]
-                              ]),
-                          /* :: */[
-                            rule("#debug nav .history", /* :: */[
-                                  /* tuple */[
-                                    "margin",
-                                    "0"
-                                  ],
-                                  /* :: */[
-                                    /* tuple */[
-                                      "padding",
-                                      "0"
-                                    ],
-                                    /* :: */[
-                                      /* tuple */[
-                                        "height",
-                                        "100%"
-                                      ],
-                                      /* :: */[
-                                        /* tuple */[
-                                          "overflow-y",
-                                          "auto"
-                                        ],
-                                        /* :: */[
-                                          /* tuple */[
-                                            "list-style",
-                                            "none"
-                                          ],
-                                          /* [] */0
-                                        ]
-                                      ]
-                                    ]
-                                  ]
-                                ]),
-                            /* :: */[
-                              rule("#debug nav .history li", /* :: */[
-                                    /* tuple */[
-                                      "margin",
-                                      "0"
-                                    ],
-                                    /* :: */[
-                                      /* tuple */[
-                                        "padding",
-                                        "0.2ch"
-                                      ],
-                                      /* [] */0
-                                    ]
-                                  ]),
-                              /* :: */[
-                                rule("#debug nav .history li.selected", /* :: */[
-                                      /* tuple */[
-                                        "background-color",
-                                        "#333"
-                                      ],
-                                      /* [] */0
-                                    ]),
-                                /* :: */[
-                                  rule("#debug nav .history span.details", /* :: */[
-                                        /* tuple */[
-                                          "display",
-                                          "inline-block"
-                                        ],
-                                        /* :: */[
-                                          /* tuple */[
-                                            "cursor",
-                                            "pointer"
-                                          ],
-                                          /* :: */[
-                                            /* tuple */[
-                                              "width",
-                                              "1ch"
-                                            ],
-                                            /* :: */[
-                                              /* tuple */[
-                                                "margin",
-                                                "0 1ch"
-                                              ],
-                                              /* :: */[
-                                                /* tuple */[
-                                                  "vertical-align",
-                                                  "super"
-                                                ],
-                                                /* [] */0
-                                              ]
-                                            ]
-                                          ]
-                                        ]
-                                      ]),
-                                  /* :: */[
-                                    rule("#debug nav .history li.selected span.details:after", /* :: */[
-                                          /* tuple */[
-                                            "content",
-                                            "'\\2026'"
-                                          ],
-                                          /* [] */0
-                                        ]),
-                                    /* :: */[
-                                      rule("#debug nav .history li.selected span.details.show:before", /* :: */[
-                                            /* tuple */[
-                                              "position",
-                                              "absolute"
-                                            ],
-                                            /* :: */[
-                                              /* tuple */[
-                                                "content",
-                                                "' '"
-                                              ],
-                                              /* :: */[
-                                                /* tuple */[
-                                                  "border",
-                                                  "solid transparent"
-                                                ],
-                                                /* :: */[
-                                                  /* tuple */[
-                                                    "border-right-color",
-                                                    "#333"
-                                                  ],
-                                                  /* :: */[
-                                                    /* tuple */[
-                                                      "border-width",
-                                                      "1.6ch"
-                                                    ],
-                                                    /* :: */[
-                                                      /* tuple */[
-                                                        "margin-left",
-                                                        "-4.4ch"
-                                                      ],
-                                                      /* :: */[
-                                                        /* tuple */[
-                                                          "margin-top",
-                                                          "-.3ch"
-                                                        ],
-                                                        /* [] */0
-                                                      ]
-                                                    ]
-                                                  ]
-                                                ]
-                                              ]
-                                            ]
-                                          ]),
-                                      /* :: */[
-                                        rule("#debug nav .history span.message", /* :: */[
-                                              /* tuple */[
-                                                "display",
-                                                "inline-block"
-                                              ],
-                                              /* :: */[
-                                                /* tuple */[
-                                                  "cursor",
-                                                  "pointer"
-                                                ],
-                                                /* :: */[
-                                                  /* tuple */[
-                                                    "white-space",
-                                                    "nowrap"
-                                                  ],
-                                                  /* :: */[
-                                                    /* tuple */[
-                                                      "overflow",
-                                                      "hidden"
-                                                    ],
-                                                    /* :: */[
-                                                      /* tuple */[
-                                                        "text-overflow",
-                                                        "ellipsis"
-                                                      ],
-                                                      /* [] */0
-                                                    ]
-                                                  ]
-                                                ]
-                                              ]
-                                            ]),
-                                        /* :: */[
-                                          rule("#debug nav .history span.index", /* :: */[
-                                                /* tuple */[
-                                                  "display",
-                                                  "inline-block"
-                                                ],
-                                                /* :: */[
-                                                  /* tuple */[
-                                                    "min-width",
-                                                    "3ch"
-                                                  ],
-                                                  /* :: */[
-                                                    /* tuple */[
-                                                      "margin",
-                                                      "0 1ch"
-                                                    ],
-                                                    /* :: */[
-                                                      /* tuple */[
-                                                        "color",
-                                                        "#aaa"
-                                                      ],
-                                                      /* :: */[
-                                                        /* tuple */[
-                                                          "text-align",
-                                                          "right"
-                                                        ],
-                                                        /* :: */[
-                                                          /* tuple */[
-                                                            "float",
-                                                            "right"
-                                                          ],
-                                                          /* [] */0
-                                                        ]
-                                                      ]
-                                                    ]
-                                                  ]
-                                                ]
-                                              ]),
-                                          /* :: */[
-                                            rule("#debug aside.details", /* :: */[
-                                                  /* tuple */[
-                                                    "position",
-                                                    "absolute"
-                                                  ],
-                                                  /* :: */[
-                                                    /* tuple */[
-                                                      "width",
-                                                      "40ch"
-                                                    ],
-                                                    /* :: */[
-                                                      /* tuple */[
-                                                        "top",
-                                                        "0"
-                                                      ],
-                                                      /* :: */[
-                                                        /* tuple */[
-                                                          "bottom",
-                                                          "0"
-                                                        ],
-                                                        /* :: */[
-                                                          /* tuple */[
-                                                            "right",
-                                                            "100%"
-                                                          ],
-                                                          /* :: */[
-                                                            /* tuple */[
-                                                              "margin-right",
-                                                              "1.5ch"
-                                                            ],
-                                                            /* :: */[
-                                                              /* tuple */[
-                                                                "overflow",
-                                                                "scroll"
-                                                              ],
-                                                              /* :: */[
-                                                                /* tuple */[
-                                                                  "background-color",
-                                                                  "#fff"
-                                                                ],
-                                                                /* :: */[
-                                                                  /* tuple */[
-                                                                    "color",
-                                                                    "#000"
-                                                                  ],
-                                                                  /* :: */[
-                                                                    /* tuple */[
-                                                                      "box-shadow",
-                                                                      "0 0 10px #333"
-                                                                    ],
-                                                                    /* :: */[
-                                                                      /* tuple */[
-                                                                        "border-radius",
-                                                                        "4px 4px 0 0"
-                                                                      ],
-                                                                      /* :: */[
-                                                                        /* tuple */[
-                                                                          "border",
-                                                                          "2px solid #333"
-                                                                        ],
-                                                                        /* :: */[
-                                                                          /* tuple */[
-                                                                            "padding",
-                                                                            "1ch"
-                                                                          ],
-                                                                          /* :: */[
-                                                                            /* tuple */[
-                                                                              "white-space",
-                                                                              "pre"
-                                                                            ],
-                                                                            /* [] */0
-                                                                          ]
-                                                                        ]
-                                                                      ]
-                                                                    ]
-                                                                  ]
-                                                                ]
-                                                              ]
-                                                            ]
-                                                          ]
-                                                        ]
-                                                      ]
-                                                    ]
-                                                  ]
-                                                ]),
-                                            /* [] */0
-                                          ]
-                                        ]
-                                      ]
-                                    ]
-                                  ]
-                                ]
-                              ]
-                            ]
-                          ]
-                        ]
-                      ]
-                    ]
-                  ]
-                ]
-              ]);
-  };
-  var view_details = function (model) {
-    var format = (
-        function (v) { return JSON.stringify(v, null, 2); }
-      );
-    return Tea_html2.aside(/* None */0, /* None */0, /* :: */[
-                Curry._1(Tea_html2.Attributes[/* class' */3], "details"),
-                /* [] */0
-              ], /* :: */[
-                /* Text */Block.__(1, [Curry._1(format, model)]),
-                /* [] */0
-              ]);
-  };
-  var view_history = function (model, selected_index) {
-    var count = List.length(model[/* history */0]);
-    return Tea_html2.ul(/* None */0, /* None */0, /* :: */[
-                Curry._1(Tea_html2.Attributes[/* class' */3], "history"),
-                /* [] */0
-              ], List.mapi((function (i, param) {
-                      var selected = +(i === selected_index);
-                      return Tea_html2.li(/* None */0, /* None */0, /* :: */[
-                                  Curry._1(Tea_html2.Events[/* onClick */8], /* SelectHistoryItem */Block.__(1, [i])),
-                                  /* :: */[
-                                    Curry._1(Tea_html2.Attributes[/* classList */4], /* :: */[
-                                          /* tuple */[
-                                            "selected",
-                                            selected
-                                          ],
-                                          /* [] */0
-                                        ]),
-                                    /* [] */0
-                                  ]
-                                ], /* :: */[
-                                  Tea_html2.span(/* None */0, /* None */0, /* :: */[
-                                        Curry._1(Tea_html2.Attributes[/* classList */4], /* :: */[
-                                              /* tuple */[
-                                                "details",
-                                                /* true */1
-                                              ],
-                                              /* :: */[
-                                                /* tuple */[
-                                                  "show",
-                                                  selected && model[/* show_details */2]
-                                                ],
-                                                /* [] */0
-                                              ]
-                                            ]),
-                                        selected ? /* :: */[
-                                            Curry._1(Tea_html2.Events[/* onClick */8], /* ToggleDetails */1),
-                                            /* :: */[
-                                              Curry._1(Tea_html2.Attributes[/* title */6], "toggle details"),
-                                              /* [] */0
-                                            ]
-                                          ] : /* :: */[
-                                            Tea_html2.Attributes[/* noProp */0],
-                                            /* :: */[
-                                              Tea_html2.Attributes[/* noProp */0],
-                                              /* [] */0
-                                            ]
-                                          ]
-                                      ], /* :: */[
-                                        selected && model[/* show_details */2] ? view_details(param[1]) : Tea_html2.noNode,
-                                        /* [] */0
-                                      ]),
-                                  /* :: */[
-                                    Tea_html2.span(/* None */0, /* None */0, /* :: */[
-                                          Curry._1(Tea_html2.Attributes[/* class' */3], "message"),
-                                          /* [] */0
-                                        ], /* :: */[
-                                          /* Text */Block.__(1, [param[0]]),
-                                          /* [] */0
-                                        ]),
-                                    /* :: */[
-                                      Tea_html2.span(/* None */0, /* None */0, /* :: */[
-                                            Curry._1(Tea_html2.Attributes[/* class' */3], "index"),
-                                            /* [] */0
-                                          ], /* :: */[
-                                            /* Text */Block.__(1, [String(count - i | 0)]),
-                                            /* [] */0
-                                          ]),
-                                      /* [] */0
-                                    ]
-                                  ]
-                                ]);
-                    }), model[/* history */0]));
-  };
-  var view$prime = function (model) {
-    var match = model[/* state */1];
-    var match$1;
-    if (match) {
-      var index = match[0];
-      match$1 = /* tuple */[
-        index,
-        List.nth(model[/* history */0], index)[1],
-        /* true */1
-      ];
-    } else {
-      match$1 = /* tuple */[
-        0,
-        List.hd(model[/* history */0])[1],
-        /* false */0
-      ];
-    }
-    var paused = match$1[2];
-    var history_count = List.length(model[/* history */0]);
-    var vnode = Curry._1(view, match$1[1]);
-    return Tea_html2.div(/* None */0, /* None */0, /* [] */0, /* :: */[
-                Vdom.map(client_msg, vnode),
-                /* :: */[
-                  Tea_html2.div(/* None */0, /* None */0, /* :: */[
-                        Curry._1(Tea_html2.Attributes[/* id */5], "debug"),
-                        /* :: */[
-                          Curry._1(Tea_html2.Attributes[/* classList */4], /* :: */[
-                                /* tuple */[
-                                  "paused",
-                                  paused
-                                ],
-                                /* [] */0
-                              ]),
-                          /* [] */0
-                        ]
-                      ], /* :: */[
-                        view_styles(/* () */0),
-                        /* :: */[
-                          Tea_html2.nav(/* None */0, /* None */0, /* [] */0, /* :: */[
-                                Tea_html2.div(/* None */0, /* None */0, /* :: */[
-                                      Curry._1(Tea_html2.Attributes[/* class' */3], "toggle"),
-                                      /* :: */[
-                                        Curry._1(Tea_html2.Events[/* onClick */8], /* TogglePaused */0),
-                                        /* :: */[
-                                          paused ? Curry._1(Tea_html2.Attributes[/* title */6], "click to resume") : Curry._1(Tea_html2.Attributes[/* title */6], "click to pause"),
-                                          /* [] */0
-                                        ]
-                                      ]
-                                    ], /* :: */[
-                                      /* Text */Block.__(1, [Curry._1(Printf.sprintf(/* Format */[
-                                                    /* String_literal */Block.__(11, [
-                                                        "Explore History (",
-                                                        /* Int */Block.__(4, [
-                                                            /* Int_d */0,
-                                                            /* No_padding */0,
-                                                            /* No_precision */0,
-                                                            /* Char_literal */Block.__(12, [
-                                                                /* ")" */41,
-                                                                /* End_of_format */0
-                                                              ])
-                                                          ])
-                                                      ]),
-                                                    "Explore History (%d)"
-                                                  ]), history_count)]),
-                                      /* [] */0
-                                    ]),
-                                /* :: */[
-                                  paused ? view_history(model, match$1[0]) : Tea_html2.noNode,
-                                  /* [] */0
-                                ]
-                              ]),
-                          /* [] */0
-                        ]
-                      ]),
-                  /* [] */0
-                ]
-              ]);
-  };
-  var subscriptions$prime = function (model) {
-    return Tea_sub.map(client_msg, Curry._1(subscriptions, List.hd(model[/* history */0])[1]));
-  };
-  return /* record */[
-          /* init */init$prime,
-          /* update */update$prime,
-          /* view */view$prime,
-          /* subscriptions */subscriptions$prime
-        ];
-}
-
-function standardProgram(options, pnode, flags) {
-  return Tea_app.standardProgram(wrap(options), pnode, flags);
-}
-
-exports.wrap = wrap;
-exports.standardProgram = standardProgram;
-/* Tea_html2 Not a pure module */
