+++ test/index.html
@@ -12,7 +12,6 @@
     <button onclick="loadApp()">Shutdown</button>
     <button onclick="loadApp('counter')">Counter</button>
     <button onclick="loadApp('counter_debug')">Counter (with debugger)</button>
-    <button onclick="loadApp('counter_debug_records')">Counter (with debugger records)</button>
     <button onclick="loadApp('btn_update_span')">Button Update Span</button>
     <button onclick="loadApp('attribute_removal')">Attribute Removal</button>
     <button onclick="loadApp('drag')">Drag</button>
