+++ test/test_client.ml
@@ -1,6 +1,5 @@
 let counter = let open Test_client_counter in main
 let counter_debug = let open Test_client_counter_debug in main
-let counter_debug_records = let open Test_client_counter_debug_records in main
 let btn_update_span = let open Test_client_btn_update_span in main
 let attribute_removal = let open Test_client_attribute_removal in main
 let drag = let open Test_client_drag in main
