--- test/test_client_counter_debug_records.ml
@@ -1,64 +1,0 @@
-open Tea.App
-open Tea.Html
-
-type msg =
-  | Increment
-  | Decrement
-  | Reset
-  | Set of int
-
-let string_of_msg = function
-  | Increment -> "Increment"
-  | Decrement -> "Decrement"
-  | Reset -> "Reset"
-  | Set _ -> "Set"
-
-let init () = 4, Tea.Cmd.none
-
-let subscriptions _ = Tea.Sub.none
-
-let update model = function
-  | Increment -> model + 1, Tea.Cmd.none
-  | Decrement -> model - 1, Tea.Cmd.none
-  | Reset -> 0, Tea.Cmd.none
-  | Set v -> v, Tea.Cmd.none
-
-
-let view_button title msg =
-  button
-    [ onClick msg
-    ]
-    [ text title
-    ]
-
-let view model =
-  div
-    []
-    [ span
-        [ style "text-weight" "bold" ]
-        [ text (string_of_int model) ]
-    ; br []
-    ; view_button "Increment" (
-        if model >= 3 then
-          Decrement
-        else
-          Increment
-      )
-    ; br []
-    ; view_button "Decrement" Decrement
-    ; br []
-    ; view_button "Set to 42" (Set 42)
-    ; br []
-    ; if model <> 0 then view_button "Reset" Reset else noNode
-    ]
-
-module Debug = Tea_debug_records
-
-let main =
-  Debug.standardProgram {
-    init;
-    update;
-    view;
-    subscriptions;
-    string_of_msg;
-  }
